# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: PIDEA Project-Centric API
  description: |
    RESTful API for project-centric development environment management.
    This API provides endpoints for managing projects and their associated interfaces,
    replacing the legacy IDE-centric API structure.
  version: 2.0.0
  contact:
    name: PIDEA Development Team
    email: dev@pidea.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.pidea.io/api
    description: Production server
...
paths:
  /projects:
    get:
      tags:
        - Projects
      summary: List projects
      description: Retrieve a paginated list of projects with optional search filtering
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of projects per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: search
          in: query
          description: Search term for filtering projects
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 100
      responses:
        '200':
          description: Successful response with project list
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
                  total:
                    type: integer
                    description: Total number of projects
                  page:
                    type: integer
                    description: Current page number
                  limit:
                    type: integer
                    description: Number of projects per page
              example:
                projects:
                  - id: "proj_1234567890"
                    name: "My Web Project"
                    workspacePath: "/path/to/project"
                    description: "A web development project"
                    type: "web"
                    framework: "react"
                    createdAt: "2025-10-11T02:20:03.000Z"
                    updatedAt: "2025-10-11T02:20:03.000Z"
                total: 1
                page: 1
                limit: 10
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Projects
      summary: Create a new project
      description: Create a new project with the specified configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreateRequest'
            example:
              name: "My New Project"
              workspacePath: "/path/to/new/project"
              description: "A new development project"
              type: "web"
              framework: "react"
      responses:
        '201':
          description: Project created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  project:
                    $ref: '#/components/schemas/Project'
              example:
                project:
                  id: "proj_1234567890"
                  name: "My New Project"
                  workspacePath: "/path/to/new/project"
                  description: "A new development project"
                  type: "web"
                  framework: "react"
                  createdAt: "2025-10-11T02:20:03.000Z"
                  updatedAt: "2025-10-11T02:20:03.000Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /projects/{projectId}:
    get:
      tags:
        - Projects
      summary: Get a project by ID
      description: Retrieve a specific project by its ID
      parameters:
        - name: projectId
          in: path
          required: true
          description: Unique identifier for the project
          schema:
            type: string
            pattern: '^proj_[a-zA-Z0-9]+$'
      responses:
        '200':
          description: Project retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  project:
                    $ref: '#/components/schemas/Project'
              example:
                project:
                  id: "proj_1234567890"
                  name: "My Web Project"
                  workspacePath: "/path/to/project"
                  description: "A web development project"
                  type: "web"
                  framework: "react"
                  createdAt: "2025-10-11T02:20:03.000Z"
                  updatedAt: "2025-10-11T02:20:03.000Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Projects
      summary: Update a project
      description: Update an existing project with new data
      parameters:
        - name: projectId
          in: path
          required: true
          description: Unique identifier for the project
          schema:
            type: string
            pattern: '^proj_[a-zA-Z0-9]+$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdateRequest'
            example:
              name: "Updated Project Name"
              description: "Updated project description"
      responses:
        '200':
          description: Project updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  project:
                    $ref: '#/components/schemas/Project'
              example:
                project:
                  id: "proj_1234567890"
                  name: "Updated Project Name"
                  workspacePath: "/path/to/project"
                  description: "Updated project description"
                  type: "web"
                  framework: "react"
                  createdAt: "2025-10-11T02:20:03.000Z"
                  updatedAt: "2025-10-11T02:21:03.000Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Projects
      summary: Delete a project
      description: Delete a project and all its associated interfaces
      parameters:
        - name: projectId
          in: path
          required: true
          description: Unique identifier for the project
          schema:
            type: string
            pattern: '^proj_[a-zA-Z0-9]+$'
      responses:
        '200':
          description: Project deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Success message
              example:
                message: "Project deleted successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /projects/{projectId}/interfaces:
    get:
      tags:
        - Interfaces
      summary: List interfaces for a project
      description: Retrieve all interfaces associated with a project
      parameters:
        - name: projectId
          in: path
          required: true
          description: Unique identifier for the project
          schema:
            type: string
            pattern: '^proj_[a-zA-Z0-9]+$'
      responses:
        '200':
          description: Interfaces retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  interfaces:
                    type: array
                    items:
                      $ref: '#/components/schemas/Interface'
              example:
                interfaces:
                  - id: "iface_1234567890"
                    name: "VS Code Interface"
                    type: "ide"
                    projectId: "proj_1234567890"
                    configuration:
                      port: 3000
                      workspacePath: "/workspace"
                    status: "running"
                    createdAt: "2025-10-11T02:20:03.000Z"
                    updatedAt: "2025-10-11T02:20:03.000Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Interfaces
      summary: Create an interface for a project
      description: Create a new interface within a project
      parameters:
        - name: projectId
          in: path
          required: true
          description: Unique identifier for the project
          schema:
            type: string
            pattern: '^proj_[a-zA-Z0-9]+$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InterfaceCreateRequest'
            example:
              name: "VS Code Interface"
              type: "ide"
              configuration:
                port: 3000
                workspacePath: "/workspace"
      responses:
        '201':
          description: Interface created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  interface:
                    $ref: '#/components/schemas/Interface'
              example:
                interface:
                  id: "iface_1234567890"
                  name: "VS Code Interface"
                  type: "ide"
                  projectId: "proj_1234567890"
                  configuration:
                    port: 3000
                    workspacePath: "/workspace"
                  status: "created"
                  createdAt: "2025-10-11T02:20:03.000Z"
                  updatedAt: "2025-10-11T02:20:03.000Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
    Project:
      type: object
      required:
        - id
        - name
        - workspacePath
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          pattern: '^proj_[a-zA-Z0-9]+$'
          description: Unique identifier for the project
          example: "proj_1234567890"
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: Project name
          example: "My Web Project"
        workspacePath:
          type: string
          minLength: 1
          maxLength: 500
          description: Path to the project workspace
          example: "/path/to/project"
        description:
          type: string
          maxLength: 1000
          description: Project description
          example: "A web development project"
        type:
          type: string
          enum: [web, mobile, desktop, api, library]
          description: Project type
          example: "web"
        framework:
          type: string
          description: Project framework
          example: "react"
        createdAt:
          type: string
          format: date-time
          description: Project creation timestamp
          example: "2025-10-11T02:20:03.000Z"
        updatedAt:
          type: string
          format: date-time
          description: Project last update timestamp
          example: "2025-10-11T02:20:03.000Z"
    ProjectCreateRequest:
      type: object
      required:
        - name
        - workspacePath
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: Project name
          example: "My New Project"
        workspacePath:
          type: string
          minLength: 1
          maxLength: 500
          description: Path to the project workspace
          example: "/path/to/new/project"
        description:
          type: string
          maxLength: 1000
          description: Project description
          example: "A new development project"
        type:
          type: string
          enum: [web, mobile, desktop, api, library]
          description: Project type
          example: "web"
        framework:
          type: string
          description: Project framework
          example: "react"
    ProjectUpdateRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: Project name
          example: "Updated Project Name"
        workspacePath:
          type: string
          minLength: 1
          maxLength: 500
          description: Path to the project workspace
          example: "/path/to/updated/project"
        description:
          type: string
          maxLength: 1000
          description: Project description
          example: "Updated project description"
        type:
          type: string
          enum: [web, mobile, desktop, api, library]
          description: Project type
          example: "web"
        framework:
          type: string
          description: Project framework
          example: "react"
    Interface:
      type: object
      required:
        - id
        - name
        - type
        - projectId
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          pattern: '^iface_[a-zA-Z0-9]+$'
          description: Unique identifier for the interface
          example: "iface_1234567890"
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: Interface name
          example: "VS Code Interface"
        type:
          type: string
          enum: [ide, editor, terminal, browser, file-system, database, api, websocket]
          description: Interface type
          example: "ide"
        projectId:
          type: string
          pattern: '^proj_[a-zA-Z0-9]+$'
          description: Project identifier
          example: "proj_1234567890"
        configuration:
          type: object
          description: Interface configuration
          example:
            port: 3000
            workspacePath: "/workspace"
        status:
          type: string
          enum: [created, starting, running, stopping, stopped, error]
          description: Interface status
          example: "running"
        createdAt:
          type: string
          format: date-time
          description: Interface creation timestamp
          example: "2025-10-11T02:20:03.000Z"
        updatedAt:
          type: string
          format: date-time
          description: Interface last update timestamp
          example: "2025-10-11T02:20:03.000Z"
    InterfaceCreateRequest:
      type: object
      required:
        - name
        - type
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: Interface name
          example: "VS Code Interface"
        type:
          type: string
          enum: [ide, editor, terminal, browser, file-system, database, api, websocket]
          description: Interface type
          example: "ide"
        configuration:
          type: object
          description: Interface configuration
          example:
            port: 3000
            workspacePath: "/workspace"
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type
          example: "ValidationError"
        message:
          type: string
          description: Error message
          example: "Invalid project ID format"
        details:
          type: object
          description: Additional error details
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2025-10-11T02:20:03.000Z"
  responses:
    BadRequest:
      description: Bad request - invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "ValidationError"
            message: "Invalid project ID format"
            timestamp: "2025-10-11T02:20:03.000Z"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "NotFoundError"
            message: "Project not found"
            timestamp: "2025-10-11T02:20:03.000Z"
    Conflict:
      description: Resource conflict - resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "ConflictError"
            message: "Project with this name already exists"
            timestamp: "2025-10-11T02:20:03.000Z"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "InternalServerError"
            message: "An unexpected error occurred"
            timestamp: "2025-10-11T02:20:03.000Z"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
security:
  - BearerAuth: []